+++ src/App.js
@@ -1,7 +1,7 @@
 import React from "react";
 import "./App.css";
 import "./components/defaultCss/style.css";
-import Menu from "./components/menu/menu";
+import Menu from "./components/menu";
 import Header from "./components/header/header";
 import Sobre from "./components/sobre/sobre";
 import Portfolio from "./components/portfolio/";
